refactor: Move BinaryLoaderSerializer class into a folder
Seperate code out of the same file and into their own files for the binary serializer interface and class
