+++ src/CoreHook.BinaryInjection/BinaryLoader/BinaryLoader.Windows.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using CoreHook.BinaryInjection.Host;
+using CoreHook.BinaryInjection.BinaryLoader.Serializer;
 using CoreHook.Memory;
 
 namespace CoreHook.BinaryInjection.BinaryLoader
