+++ src/CoreHook.BinaryInjection/BinaryLoader/Serializer/BinaryLoaderSerializer.cs
@@ -1,35 +1,7 @@
 ﻿using System.IO;
-using System.Text;
-using CoreHook.BinaryInjection.Host;
 
-namespace CoreHook.BinaryInjection.BinaryLoader
+namespace CoreHook.BinaryInjection.BinaryLoader.Serializer
 {
-    public interface IBinaryLoaderConfig
-    {
-        int MaxPathLength { get; }
-        Encoding PathEncoding { get; }
-
-    }
-    public interface IBinaryLoaderHostConfig
-    {
-        string CoreCLRStartFunction { get; }
-        string CoreCLRExecuteManagedFunction { get; }
-    }
-
-    public sealed partial class BinaryLoaderHostConfig : IBinaryLoaderHostConfig
-    {
-        public string CoreCLRStartFunction
-            => "StartCoreCLR";
-        public string CoreCLRExecuteManagedFunction
-            => "ExecuteAssemblyFunction";
-    }
-
-    public sealed partial class BinaryLoaderConfig : IBinaryLoaderConfig
-    {
-        public int MaxPathLength => 260; 
-        public Encoding PathEncoding => Encoding.Unicode;
-    }
-
     public class BinaryLoaderSerializer : IBinarySerializer
     {
         public BinaryLoaderArguments Arguments { get; set; }
@@ -60,5 +32,4 @@ namespace CoreHook.BinaryInjection.BinaryLoader
             }
         }
     }
-
 }
