+++ src/CoreHook.BinaryInjection/BinaryLoader/Serializer/IBinarySerializer.cs
@@ -1,0 +1,13 @@
+﻿using System.IO;
+
+namespace CoreHook.BinaryInjection.BinaryLoader.Serializer
+{
+    public interface IBinarySerializer
+    {
+        byte[] Serialize();
+    }
+    public interface IBinaryStreamSerializer
+    {
+        void Serialize(MemoryStream stream);
+    }
+}
